+++ include/fpag/mem/page_allocator.h
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <cstddef>
 #include <cstdint>
 
 #include "fpag/base/numeric.h"
@@ -36,8 +37,9 @@ void decommit_pages(void* ptr, usize size);
 // `size` is not used on Windows, but is required on POSIX systems.
 void free_pages(void* ptr, usize size);
 
-static constexpr u64 kPageSize = 4096;                 // 4 KiB
-static constexpr u64 kHugePageSize = 2 * 1024 * 1024;  // 2 MiB
+static constexpr u64 kPageSize = 4096;  // 4 KiB
+static constexpr u64 kHugePageSize =
+    static_cast<const u64>(2 * 1024 * 1024);  // 2 MiB
 
 inline bool is_page_aligned_ptr(void* ptr) {
   return reinterpret_cast<uintptr_t>(ptr) % kPageSize == 0;
